requirements.txt
streamlit>=1.28
Pillow>=9.0
numpy>=1.24
scipy>=1.10
opencv-python-headless>=4.7
requests>=2.28
piexif>=1.1.3
PyWavelets>=1.4
scikit-image>=0.20
torch
torchvision
